Move functionally from RentalInfo into MovieRental
adds movie to movieRental instead of movie id.

In the MovieRental constructor it takes movieName to create movie. could in theory be the old movie id, but I think just using movie name makes the code clearer and easier to understand.

Doing this makes the app easier to expand with new movies since they only have to be added to they MovieRental constructor switch case.

since movie got accessible from MovieRental it is possible to have calculatePrice in there instead of in rentalInfo. which makes it easier to add new movie codes and expand the application in the future.

due to these changes it was possible to remove the hashmap and price calculations from RentalInfo